requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-genai==0.5.0
pydantic==2.5.0
python-multipart==0.0.6